requirements.txt
torch==2.4.0
torchaudio==2.4.0
tokenizers==0.21.0
transformers==4.49.0
huggingface_hub==0.28.1
moshi==0.2.2
torchtune==0.4.0
torchao==0.9.0
silentcipher @ git+https://github.com/SesameAILabs/silentcipher@master
hf_transfer==0.1.5
soundfile>=0.12.1
# Including both Gemini API packages to ensure compatibility
google-genai>=0.1.7
google-generativeai>=0.8.1
python-dotenv>=1.0.0
protobuf>=4.25.0
